Add documentation generator for extensions
This commit adds docgen.js, a documentation generator for compiled MIT
App Inventor extensions. This script runs with phantomjs and given a
.aix file will generate one Markdown file per extension present in the
.aix as well as a blocks directory with Scalable Vector Graphics (SVG)
images for each block (getter, setter, method, event) defined by the
extension. Content for the Markdown file is taken from the description
fields of the extension attributes, which are either set in the
associated Simple annotation or from Javadoc style comment in the
extension's source code.

Change-Id: I5e2fb5d7cfe087cd12009887ef21fccfd70243e6